+++ starsky/starsky/clientapp/src/components/detail-view-sidebar.tsx
@@ -2,6 +2,7 @@ import { Link } from '@reach/router';
 import React, { memo, useEffect, useRef } from "react";
 import { useDetailViewContext } from '../contexts/detailview-context';
 import useFetch from '../hooks/use-fetch';
+import useGlobalSettings from '../hooks/use-global-settings';
 import useKeyboardEvent from '../hooks/use-keyboard-event';
 import useLocation from '../hooks/use-location';
 import { IExifStatus } from '../interfaces/IExifStatus';
@@ -11,6 +12,7 @@ import { CastToInterface } from '../shared/cast-to-interface';
 import { isValidDate, parseDate, parseRelativeDate, parseTime } from '../shared/date';
 import FetchPost from '../shared/fetch-post';
 import { Keyboard } from '../shared/keyboard';
+import { Language } from '../shared/language';
 import { URLPath } from '../shared/url-path';
 import { UrlQuery } from '../shared/url-query';
 import ColorClassSelect from './color-class-select';
@@ -20,6 +22,24 @@ interface IDetailViewSidebarProps {
 }
 
 const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo((props) => {
+
+  // content
+  const settings = useGlobalSettings();
+  const language = new Language(settings.language);
+  const MessageTitleName = language.text("Titel", "Title");
+  const MessageInfoName = "Info";
+  const MessageColorClassification = language.text("Kleur-Classificatie", "Color Classification")
+  const MessageDateTimeAgoEdited = language.text("geleden bewerkt", "ago edited");
+  const MessageDateLessThan1Minute = language.text("minder dan één minuut", "less than one minute");
+  const MessageDateMinutes = language.text("minuten", "minutes");
+  const MessageDateHour = language.text("uur", "less than one minute");
+  const MessageNounUnknown = language.text("Onbekende", "Unknown");
+  const MessageLocation = language.text("locatie", "location");
+  const MessageReadOnlyFile = language.text("Alleen lezen bestand", "Read only file");
+  const MessageNotFoundSourceMissing = language.text("Mist in de index", "Misses in the index");
+  const MessageServerError = language.text("Er is iets mis met de input", "Something is wrong with the input");
+  const MessageDeleted = language.text("Staat in de prullenmand", "Is in the trash");
+  const MessageDeletedRestoreInstruction = language.text("'Zet terug uit prullenmand' om het item te bewerken", "'Restore from Trash' to edit the item");
 
   let { state, dispatch } = useDetailViewContext();
   var history = useLocation();
@@ -75,7 +95,7 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
 
 
   function handleChange(event: React.ChangeEvent<HTMLDivElement>) {
-    let value = event.currentTarget.innerText;
+    let value = event.currentTarget.textContent;
     let name = event.currentTarget.dataset["name"];
 
     if (!name) return;
@@ -83,32 +103,30 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
 
     // compare
     var fileIndexObject: any = fileIndexItem;
+
     if (!fileIndexObject[name] === undefined) return; //to update emthy start to first fill
 
     var currentString: string = fileIndexObject[name];
     if (value === currentString) return;
 
-    // Empty strings are NOT supported
-    if (event.currentTarget.innerText.length === 1) {
-      fileIndexObject[name] = ".";
-      console.log('not supported');
-    }
-
     var updateObject: any = { f: fileIndexItem.filePath };
-    updateObject[name] = value;
+    updateObject[name] = value.trim();
 
-    var updateApiUrl = new UrlQuery().UrlUpdateApi();
     var bodyParams = new URLPath().ObjectToSearchParams(updateObject);
 
-    FetchPost(updateApiUrl, bodyParams.toString()).then(item => {
-      if (item.statusCode !== 200) return;
+    FetchPost(new UrlQuery().UrlUpdateApi(), bodyParams.toString()).then(item => {
+      if (item.statusCode !== 200 || !item.data) return;
 
       var currentItem = item.data[0] as IFileIndexItem;
       currentItem.lastEdited = new Date().toISOString();
       setFileIndexItem(currentItem);
       dispatch({ 'type': 'update', ...currentItem });
 
-      // clear search cache new URLPath().StringToIUrl(history.location.search).t
+      // clear search cache
+      var searchTag = new URLPath().StringToIUrl(history.location.search).t;
+
+      if (!searchTag) return;
+      FetchPost(new UrlQuery().UrlSearchRemoveCacheApi(), `t=${searchTag}`);
     });
   }
 
@@ -124,14 +142,18 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
   // noinspection HtmlUnknownAttribute
   return (<div className="sidebar">
     {fileIndexItem.status === IExifStatus.Deleted || fileIndexItem.status === IExifStatus.ReadOnly
-      || fileIndexItem.status === IExifStatus.NotFoundSourceMissing || fileIndexItem.status === IExifStatus.ServerError ? <><div className="content--header">
-        Status
-    </div> <div className="content content--text">
-          {fileIndexItem.status === IExifStatus.Deleted ? <><div className="warning-box">Staat in de prullenmand </div> 'Zet terug uit prullenmand' om het item te bewerken</> : null}
-          {fileIndexItem.status === IExifStatus.NotFoundSourceMissing ? <><div className="warning-box">Mist in de index </div> </> : null}
-          {fileIndexItem.status === IExifStatus.ReadOnly ? <><div className="warning-box">Alleen lezen bestand</div> </> : null}
-          {fileIndexItem.status === IExifStatus.ServerError ? <><div className="warning-box">Er is iets mis met de input</div> </> : null}
-        </div></> : null}
+      || fileIndexItem.status === IExifStatus.NotFoundSourceMissing || fileIndexItem.status === IExifStatus.ServerError ? <>
+        <div className="content--header">
+          Status
+        </div>
+        <div className="content content--text">
+          {fileIndexItem.status === IExifStatus.Deleted ? <><div className="warning-box">{MessageDeleted}</div>
+            {MessageDeletedRestoreInstruction}</> : null}
+          {fileIndexItem.status === IExifStatus.NotFoundSourceMissing ? <><div className="warning-box">{MessageNotFoundSourceMissing}</div> </> : null}
+          {fileIndexItem.status === IExifStatus.ReadOnly ? <><div className="warning-box">{MessageReadOnlyFile}</div> </> : null}
+          {fileIndexItem.status === IExifStatus.ServerError ? <><div className="warning-box">{MessageServerError}</div> </> : null}
+        </div>
+      </> : null}
     <div className="content--header">
       Tags
     </div>
@@ -147,10 +169,10 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
     </div>
 
     <div className="content--header">
-      Info &amp; Titel
+      {MessageInfoName} &amp; {MessageTitleName}
     </div>
     <div className="content--text">
-      <h4>Info</h4>
+      <h4>{MessageInfoName}</h4>
       <div onBlur={handleChange}
         data-name="description"
         suppressContentEditableWarning={true}
@@ -158,7 +180,7 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
         className={isFormEnabled ? "form-control" : "form-control disabled"}>
         {fileIndexItem.description}
       </div>
-      <h4>Titel</h4>
+      <h4>{MessageTitleName}</h4>
       <div onBlur={handleChange}
         data-name="title"
         suppressContentEditableWarning={true}
@@ -169,8 +191,8 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
     </div>
 
     <div className="content--header">
-      Kleur-Classificatie
-      </div>
+      {MessageColorClassification}
+    </div>
     <div className="content--text">
       <ColorClassSelect onToggle={() => { }} filePath={fileIndexItem.filePath}
         currentColorClass={fileIndexItem.colorClass} isEnabled={isFormEnabled} />
@@ -193,8 +215,13 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
       {isValidDate(fileIndexItem.lastEdited) ?
         <div className="box" data-test="lastEdited">
           <div className="icon icon--last-edited"></div>
-          <b>{parseRelativeDate(fileIndexItem.lastEdited)}</b>
-          <p>geleden bewerkt</p>
+          <b>{
+            language.token(parseRelativeDate(
+              fileIndexItem.lastEdited),
+              ["{lessThan1Minute}", "{minutes}", "{hour}"],
+              [MessageDateLessThan1Minute, MessageDateMinutes, MessageDateHour])
+          }</b>
+          <p>{MessageDateTimeAgoEdited}</p>
         </div> : ""}
 
       {fileIndexItem.make && fileIndexItem.model && fileIndexItem.aperture && fileIndexItem.focalLength ?
@@ -223,8 +250,8 @@ const DetailViewSidebar: React.FunctionComponent<IDetailViewSidebarProps> = memo
               <b>{fileIndexItem.locationCity}</b>
               <p>{fileIndexItem.locationCountry}</p>
             </> : <>
-              <b>Onbekende </b>
-              <p>locatie</p>
+              <b>{MessageNounUnknown}</b>
+              <p>{MessageLocation}</p>
             </>}
 
         </a> : ""}