+++ starsky/starsky/clientapp/src/components/item-list-view.tsx
@@ -1,5 +1,6 @@
 import React, { memo, useEffect } from 'react';
 import useLocation from '../hooks/use-location';
+import { PageType } from '../interfaces/IDetailView';
 import { IFileIndexItem } from '../interfaces/IFileIndexItem';
 import { INavigateState } from '../interfaces/INavigateState';
 import ListImageBox from './list-image-box';
@@ -7,6 +8,7 @@ import ListImageBox from './list-image-box';
 interface ItemListProps {
   fileIndexItems: Array<IFileIndexItem>,
   colorClassUsage: Array<number>,
+  pageType: PageType;
 }
 /**
  * A list with links to the items
@@ -25,8 +27,8 @@ const ItemListView: React.FunctionComponent<ItemListProps> = memo((props) => {
 
     // for the DOM delay
     setTimeout(() => {
-      var query = '[data-filepath="' + navigationState.filePath + '"]';
-      var elementList = document.querySelectorAll(query);
+      var dataTagQuery = `[data-filepath="${navigationState.filePath}"]`
+      var elementList = document.querySelectorAll(dataTagQuery);
       if (elementList.length !== 1) return;
 
       window.scrollTo({
@@ -44,7 +46,14 @@ const ItemListView: React.FunctionComponent<ItemListProps> = memo((props) => {
 
   return (
     <div className="folder" ref={folderRef} >
-      {items.length === 0 ? props.colorClassUsage.length >= 1 ? <div className="warning-box warning-box--left">Er zijn meer items, maar deze vallen buiten je filters</div> : <div className="warning-box"> Er zijn geen foto's in deze map</div> : ""}
+      {props.pageType !== PageType.Loading ?
+        items.length === 0 ? props.colorClassUsage.length >= 1 ?
+          <div className="warning-box warning-box--left">
+            Er zijn meer items, maar deze vallen buiten je filters
+        </div> : <div className="warning-box">
+            Er zijn geen foto's in deze map
+          </div> : null : null
+      }
       {
         items.map((item, index) => (
           <ListImageBox item={item} key={item.fileName + item.lastEdited}></ListImageBox>